Project Blue Sky : Lighting Demo
-- Extended the Blue Sky demo app to support a basic lighting control demo.
When the lighting demo feature is enabled, one device can be designated (at
compile time) to be a light controller, while a second device serves as a
remote light switch.  A button press on the switch device results in it
sending a SetLogicalCircuitState command request (part of the Nest
LogicalCircuitControlTrait) to the light controller device, which changes
the state of an output pin.

The SetLogicalCircuitState command is sent as a one-way WDM command,
addressed to the unicast WiFi fabric address of the light controller device.
 The request is sent with WRM acknowledgement requested and is retransmitted
every 100ms up to 4 times unless an ACK is received from the controller.

The light controller publishes the LogicalCircuitStateTrait containing the
current state of the light, although nothing in the current version of the
demo consumes this information.