Project Blue Sky : Lighting Demo Enhancement
-- Modified the switch-side of the lighting demo app to support dimming.  The light switch
now uses two buttons, ON and OFF, to control the state and level of the light.  Quickly
pressing and releasing the ON/OFF buttons turns the light on and off.  Pressing and holding
the buttons increases/decreases the light level.

On the M5Stack, ON is mapped to the left button and OFF is mapped to the middle button.